Godot 4.3 no longer has the internal editor toast API (#62)
In Godot 4.2 and earlier, there was an internal undocumented editor
toast API that allowed to show a toast message to the user. Such a
message was used to indicate that editing rust scripts from inside the
Godot editor is not available. Unfortunately, this API was removed in
Godot 4.3, but then was turned in a proper public API with Godot 4.4.

For now, the toast message has been replaced with an error log message.
Once gdext support for the Godot 4.4 API is released, we will show a
toast from 4.4 onwards.