Refactor Script Metadata (#87)
All `RustScript*Info` types can actually be removed in favor of directly
going from `RustScript*Desc` to engine types. The info types are
remnants of the POC.

This also triggered a refactor of the `RustScriptMethodDesc`
registration to reduce the complexity. With the added builders, it's now
much cleaner to add all methods and track their ID / index.